Added automatic gx temp-folder clean-up on closing.
Old temp folders may remain should a user open a file in the temporary
folder and not release it.  We manage this by removing file _gx_warning
file, which results in an attempt to remove that
folder the next time the gx context is lost from a gxpy-based workflow.
It is always a best-effort removal.  If a folder contains a file
_gx_warning, we do NOT remove the temporary folder as this implies the
folder belongs to a running process.